refactor: move hooks to `dev.slne.surf.lobby.hook` package and enable settings integration
- Relocated various hook classes (e.g., `ParkourHook`, `TrophyHook`) to `dev.slne.surf.lobby.hook` for improved structure.
- Added `SettingsHook` to integrate with `surf-settings` API.
- Updated `InventoryInteractListener` to utilize scroll sound settings.
