+++ build.gradle.kts
@@ -17,6 +17,7 @@ dependencies {
     compileOnly("dev.slne.surf.parkour:surf-parkour-api:1.21.11-3.1.2-SNAPSHOT")
     compileOnly("dev.slne.surf.trophy:surf-trophy-api:1.21.11-1.0.0-SNAPSHOT")
     compileOnly("dev.slne.surf.profile:surf-profile-api:1.21.11-1.0.0-SNAPSHOT")
+    compileOnly("dev.slne.surf.settings:surf-settings-api:1.21.11-2.0.0-SNAPSHOT")
 }
 
 version = findProperty("version") as String
@@ -38,5 +39,6 @@ surfPaperPluginApi {
         registerSoft("Nexo")
         registerSoft("surf-trophy-paper")
         registerSoft("surf-profile-paper")
+        registerSoft("surf-settings-paper")
     }
 }
